Validated the ontology
Used Pellet engine to validate the ontology.
