+++ bcf-ontology/src/main/resources/BcfSample.ttl
@@ -1,0 +1,154 @@
+@prefix inst:  <http://example.org/> .
+@prefix owl:   <http://www.w3.org/2002/07/owl#> .
+@prefix bot:   <https://w3id.org/bot#> .
+@prefix bcfOWL: <http://lbd.arch.rwth-aachen.de/bcf/v2.1/> .
+@prefix rdfs:  <http://www.w3.org/2000/01/rdf-schema#> .
+@prefix foaf: <http://xmlns.com/foaf/0.1/> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+
+inst:project_5a9bc49e-da48-9615-bcc9-0c45d0a3ede4 a bcfOWL:Project ;
+		bcfOWL:hasName						"Reiff" .
+
+inst:ReiffExtension a bcfOWL:Extension ;
+		bcfOWL:hasProject					inst:project_5a9bc49e-da48-9615-bcc9-0c45d0a3ede4 ;
+		bcfOWL:hasTopicType					inst:Information, inst:Error ;
+		bcfOWL:hasTopicStatus				inst:Open, inst:Closed, inst:Resolved ;
+		bcfOWL:hasLabels					inst:Documentation, inst:Strutucal, inst:MEP , inst:Architecture ;
+		bcfOWL:hasPriority					inst:Low, inst:Medium, inst:High ;
+		bcfOWL:hasUser						inst:JyrkiOraskari, inst:OliverSchulz ;
+		bcfOWL:hasStage						inst:Documentation, inst:ConstructionStart, inst:ConstructionEnd .
+		
+inst:JyrkiOraskari	a	foaf:Person ;
+		foaf:firstName "Jyrki" ;
+		foaf:lastName "Oraskari" ;
+		foaf:mbox <mailto:jyrki.oraskari@dc.rwth-aachen.de>	.
+			
+inst:OliverSchulz	a	foaf:Person ;
+		foaf:firstName "Oliver" ;
+		foaf:lastName "Schulz" ;
+		foaf:mbox <mailto:schulz@dc.rwth-aachen.de>	.
+		
+inst:Information 	a 	bcfOWL:TopicType .
+
+inst:Error			a 	bcfOWL:TopicType .
+
+inst:Open			a 	bcfOWL:TopicStatus .
+
+inst:Closed			a 	bcfOWL:TopicStatus .
+
+inst:Resolved 		a 	bcfOWL:TopicStatus .
+
+inst:Documentation	a 	bcfOWL:Label .
+
+inst:Architecture 	a 	bcfOWL:Label .
+
+inst:Structural		a	bcfOWL:Label .
+
+inst:MEP			a	bcfOWL:Label .
+
+inst:Low			a	bcfOWL:Priority .
+
+inst:Medium			a	bcfOWL:Priority .
+
+inst:High			a	bcfOWL:Priority .
+
+inst:DocumentationStage	a	bcfOWL:Stage .
+
+inst:ContructionStart	a	bcfOWL:Stage .
+
+inst:ConstructionEnd	a	bcfOWL:Stage .
+
+inst:ExampleCADApplication a bcfOWL:OriginatingSystem .
+
+
+
+inst:topic_2ef7f882-8b52-4860-af69-c61dc8408e60		a	bcfOWL:Topic ;
+		bcfOWL:hasProject					inst:project_5a9bc49e-da48-9615-bcc9-0c45d0a3ede4 ;
+		bcfOWL:hasTopicStatus				inst:Closed ;
+		bcfOWL:hasTopicType					inst:Information ;
+		bcfOWL:hasTitle						"Working" ;
+		bcfOWL:hasCreationAuthor			inst:OliverSchulz ;
+		bcfOWL:hasCreationDay				"2021-01-25T17:18:50+00:00"^^xsd:dateTime ;
+		bcfOWL:hasAssignedTo				inst:OliverSchulz ;
+		bcfOWL:hasLabels					inst:Documentation ;
+		bcfOWL:hasIndex						1 ;
+		bcfOWL:hasViewpoint					inst:viewpoint_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasComment					inst:comment_66556fb6-dc5c-4ba3-9005-1751e8c0881a .
+		
+inst:comment_66556fb6-dc5c-4ba3-9005-1751e8c0881a	a bcfOWL:Comment ;
+		bcfOWL:hasProject					inst:project_5a9bc49e-da48-9615-bcc9-0c45d0a3ede4 ;
+		bcfOWL:hasCommentText					"Test Comment" ;
+		bcfOWL:hasAuthor					inst:OliverSchulz ;
+		bcfOWL:hasModifiedAuthor			inst:OliverSchulz ;
+		bcfOWL:hasCommentDate				"2021-01-25T17:18:50+00:00"^^xsd:dateTime ;
+		bcfOWL:hasViewpoint					inst:viewpoint_daca6c77-555e-4854-9e13-90668f1f41a5 .
+		
+inst:viewpoint_daca6c77-555e-4854-9e13-90668f1f41a5	a bcfOWL:Viewpoint ;
+		bcfOWL:hasProject					inst:project_5a9bc49e-da48-9615-bcc9-0c45d0a3ede4 ;
+		bcfOWL:hasPerspectiveCamera			inst:PerspectiveCamera_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasComponents				inst:Components_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasSnapshot					inst:Snapshot_daca6c77-555e-4854-9e13-90668f1f41a5 .
+			
+inst:PerspectiveCamera_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:PerspectiveCamera ;
+		bcfOWL:hasCameraViewPoint			inst:CameraViewPoint_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasCameraDirection			inst:CameraDirection_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasCameraUpVector			inst:CameraUpVector_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasFieldOfView				"30" .
+		
+inst:Snapshot_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:Snapshot ;
+		bcfOWL:hasSnapshotType				"png" ;
+		bcfOWL:hasSnapshotURI				"http://lbd.arch.rwth-aachen.de/bcf/v2.1/snapshot/daca6c77-555e-4854-9e13-90668f1f41a5" .
+
+inst:CameraViewPoint_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:CameraViewPoint ;
+		bcfOWL:hasXValue					"10.722208" ;
+		bcfOWL:hasYValue					"16.162144" ;
+		bcfOWL:hasZValue					"1.266495" .
+
+inst:CameraDirection_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:CameraDirection ;
+		bcfOWL:hasXValue					"0.9665" ;
+		bcfOWL:hasYValue					"-0.215124" ; 
+		bcfOWL:hasZValue					"-0.140001" .
+
+inst:CameraUpVector_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:CameraUpVector ;
+		bcfOWL:hasXValue					"0.136231" ;
+		bcfOWL:hasYValue					"-0.032326" ;
+		bcfOWL:hasZValue					"0.990149" .
+		
+inst:Components_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:Components ;
+		bcfOWL:hasSelection					inst:Selection_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasVisibility				inst:Visibility_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasException					inst:Exception_daca6c77-555e-4854-9e13-90668f1f41a5 .
+		
+inst:ViewSetupHints_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:ViewSetupHints ;
+		bcfOWL:SpacesVisible				true ;
+		bcfOWL:SpaceBoundariesVisible		false ;
+		bcfOWL:OpeningsVisible				true .
+		
+inst:Selection_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:Selection ;
+		bcfOWL:hasElement					inst:Element_1ZwJHS85D3YQG5AK5ER1Wc ;
+		bcfOWL:hasElement					inst:Element_1zMXMGg3H1MOhxQ9qjUS4B ;
+		bcfOWL:hasElement					inst:Element_1fJWCTBjX5EvoBol6ESMBJ .
+		
+inst:Visibility_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:Visibility ;
+		bcfOWL:hasDefaulVisibility			false ;
+		bcfOWL:hasException					inst:Exception_daca6c77-555e-4854-9e13-90668f1f41a5 ;
+		bcfOWL:hasViewSetupHints			inst:ViewSetupHints_daca6c77-555e-4854-9e13-90668f1f41a5 .
+
+inst:Exception_daca6c77-555e-4854-9e13-90668f1f41a5 a bcfOWL:Exception ;
+		bcfOWL:hasElement					inst:Element_4ScshxZO9AJBebsniSz9Yk .
+
+inst:Element_1ZwJHS85D3YQG5AK5ER1Wc a bcfOWL:Element ;
+		bcfOWL:hasIfcGuid					"1ZwJH$85D3YQG5AK5ER1Wc" ;
+		bcfOWL:hasOriginatingSystem			inst:ExampleCADApplication .
+
+inst:Element_1zMXMGg3H1MOhxQ9qjUS4B a bcfOWL:Element ;
+		bcfOWL:hasIfcGuid					"1zMXMGg3H1MOhxQ9qjUS4B" ;
+		bcfOWL:hasOriginatingSystem			inst:ExampleCADApplication .
+		
+inst:Element_1fJWCTBjX5EvoBol6ESMBJ a bcfOWL:Element ;
+		bcfOWL:hasIfcGuid					"1fJWCTBjX5EvoBol6E$MBJ" ;
+		bcfOWL:hasOriginatingSystem			inst:ExampleCADApplication .
+		
+inst:Element_4ScshxZO9AJBebsniSz9Yk a bcfOWL:Element ;
+		bcfOWL:hasIfcGuid					"4$cshxZO9AJBebsni$z9Yk" ;
+		bcfOWL:hasOriginatingSystem			inst:ExampleCADApplication.
